% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.trendfilter.R
\name{bootstrap.trendfilter}
\alias{bootstrap.trendfilter}
\title{Bootstrap the optimized trend filtering estimator to obtain variability bands}
\usage{
bootstrap.trendfilter(
  obj,
  alpha = 0.05,
  B = 100L,
  bootstrap.algorithm = c("nonparametric", "parametric", "wild"),
  return.full.ensemble = FALSE,
  prune = TRUE,
  mc.cores = detectCores()
)
}
\arguments{
\item{obj}{An object of class '\link{SURE.trendfilter}'.}

\item{alpha}{Specifies the width of the \code{1-alpha} pointwise variability
bands. Defaults to \code{alpha = 0.05}.}

\item{B}{The number of bootstrap samples used to estimate the pointwise
variability bands. Defaults to \code{B = 100}. Increase this for more precise
bands (e.g. for the final analysis you intend to publish).}

\item{return.full.ensemble}{Logical. If \code{TRUE}, the full trend filtering
bootstrap ensemble is returned as an \mjeqn{n \times B}{ascii} matrix, less
any columns potentially pruned post-hoc (see \code{prune} below). Defaults to
\code{return.full.ensemble = FALSE}.}

\item{prune}{Logical. If \code{TRUE}, then the trend filtering bootstrap
ensemble is examined for rare instances in which the optimization has
stopped at zero knots (most likely erroneously), and removes them from the
ensemble. Defaults to \code{TRUE}. Do not change this unless you know what you are
doing!}

\item{mc.cores}{Parallel computing: The number of cores to utilize. Defaults
to the number of cores detected.}
}
\value{
An object of class 'bootstrap.trendfilter'. This is a comprehensive
list containing all of the analysis' important information, data, and
results:
\item{x.eval}{(Inherited from \code{obj}) The grid of inputs the trend
filtering estimate and variability bands were evaluated on.}
\item{tf.estimate}{The trend filtering estimate of the signal, evaluated on
\code{x.eval}.}
\item{tf.standard.errors}{The standard errors of the optimized trend
filtering point estimator.}
\item{bootstrap.lower.band}{Vector of lower bounds for the
\code{1-alpha} pointwise variability band, evaluated on \code{x.eval}.}
\item{bootstrap.upper.band}{Vector of upper bounds for the
\code{1-alpha} pointwise variability band, evaluated on \code{x.eval}.}
\item{bootstrap.algorithm}{The string specifying the bootstrap algorithms
that was used. Here, always "parametric".}
\item{alpha}{The 'level' of the variability bands, i.e. \code{alpha}
produces a \code{100*(1-alpha)}\% pointwise variability band.}
\item{B}{The number of bootstrap samples used to estimate the pointwise
variability bands.}
\item{tf.bootstrap.ensemble}{(Optional) If \code{return.full.ensemble = TRUE}, the
full trend filtering bootstrap ensemble as an \mjeqn{n \times B}{ascii}
matrix, less any columns potentially pruned post-hoc (if \code{prune = TRUE}).
If \code{return.full.ensemble = FALSE}, then this will return \code{NULL}.}
\item{edf.boots}{An integer vector of the estimated number of effective
degrees of freedom of each trend filtering bootstrap estimate. These should
all be relatively close to \code{edf.min} (below).}
\item{prune}{Logical. If \code{TRUE}, then the trend filtering bootstrap
ensemble is examined for rare instances in which the optimization has
stopped at zero knots (most likely erroneously), and removes them from the
ensemble.}
\item{n.pruned}{The number of badly-converged bootstrap trend filtering
estimates pruned from the ensemble.}
\item{x}{(Inherited from \code{obj}) The vector of the observed inputs.}
\item{y}{(Inherited from \code{obj}) The vector of the observed outputs.}
\item{weights}{(Inherited from \code{obj}) A vector of weights for the observed
outputs. These are defined as \code{weights = 1 / sigma^2}, where \code{sigma} is a
vector of standard errors of the uncertainty in the output measurements.}
\item{residuals}{(Inherited from \code{obj}) \code{residuals = y - fitted.values}}
\item{k}{(Inherited from \code{obj}) The degree of the trend filtering estimator.}
\item{gammas}{(Inherited from \code{obj}) Vector of hyperparameter values tested
during validation.}
\item{gammas.min}{(Inherited from \code{obj}) Hyperparameter value that minimizes
the validation error curve.}
\item{edf}{(Inherited from \code{obj}) Integer vector of effective degrees of
freedom for trend filtering estimators fit during validation.}
\item{edf.min}{(Inherited from \code{obj}) The effective degrees of freedom of the
optimally-tuned trend filtering estimator.}
\item{i.min}{(Inherited from \code{obj}) The index of \code{gammas} that minimizes the
validation error.}
\item{validation.method}{One of \code{c("SURE", paste0(V,"-fold CV"))}.}
\item{error}{(Inherited from \code{obj}) Vector of hyperparameter validation
errors, inherited from \code{obj} (an object of class 'SURE.trendfilter').}
\item{optimization.params}{(Inherited from \code{obj}) a list of parameters that
control the trend filtering convex optimization.}
\item{n.iter}{(Inherited from \code{obj}) Vector of the number of iterations
needed for the ADMM algorithm to converge within the given tolerance, for
each hyperparameter value. If many of these are exactly equal to \code{max_iter},
then their solutions have not converged with the tolerance specified by
\code{obj_tol}. In which case, it is often prudent to increase \code{max_iter}.}
\item{n.iter.boots}{Vector of the number of iterations needed for the ADMM
algorithm to converge within the given tolerance, for each bootstrap trend
filtering estimate.}
\item{x.scale, y.scale, data.scaled}{For internal use.}
}
\description{
\loadmathjax \code{bootstrap.trendfilter} implements...
}
\details{
See
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Politsch et al. (2020a)} for the full parametric bootstrap algorithm.
}
\references{
\strong{Companion references}
\enumerate{
\item{\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Politsch et al. (2020a). Trend filtering – I. A modern statistical tool
for time-domain astronomy and astronomical spectroscopy. \emph{Monthly
Notices of the Royal Astronomical Society}, 492(3), p. 4005-4018.}} \cr
\item{\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical
signals with varying degrees of smoothness. \emph{Monthly Notices of the
Royal Astronomical Society}, 492(3), p. 4019-4032.}}}
}
\seealso{
\code{\link{SURE.trendfilter}}
}
\author{
\subsection{\cr
\strong{Collin A. Politsch, Ph.D.}}{

Email: collinpolitsch@gmail.com \cr
Website: \href{https://collinpolitsch.com/}{collinpolitsch.com} \cr
GitHub: \href{https://github.com/capolitsch/}{github.com/capolitsch} \cr \cr
}
}
