% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{coef.trendfilter}
\alias{coef.trendfilter}
\alias{coef.cv_trendfilter}
\alias{coef.sure_trendfilter}
\alias{coef.bootstrap_trendfilter}
\alias{coefficients.cv_trendfilter}
\alias{coefficients.sure_trendfilter}
\alias{coefficients.bootstrap_trendfilter}
\title{Get coefficients from a trendfilter object}
\usage{
\method{coef}{trendfilter}(obj, lambda = NULL, ...)
}
\arguments{
\item{obj}{Object of class/subclass '\code{\link[=trendfilter]{trendfilter}}'.}

\item{lambda}{One of more hyperparameter values to calculate model coefficients for.
Defaults to \code{lambda = NULL}, in which case, model coefficients are computed
for every hyperparameter value in \code{obj$lambda}.}
}
\description{
Get coefficients from a trendfilter object
}
