% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eclipsing_binary}
\alias{eclipsing_binary}
\title{Phase-folded light curve of an eclipsing binary star system}
\format{
A tibble with column set \code{c("phase", "flux", "std_err")}.
}
\source{
\url{http://keplerebs.villanova.edu/}
}
\usage{
eclipsing_binary
}
\description{
A binary star system's apparent brightness time series, collected by NASA's
\href{https://www.nasa.gov/mission_pages/kepler/overview/index.html}{\emph{Kepler} mission}. The
gravitationally-bound pair of stars in the system orbit each other with a
period of ~1.56 Earth days. The system's \emph{Kepler} identification is KIC
6048106.
}
\keyword{datasets}
