% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{fitted.trendfilter}
\alias{fitted.trendfilter}
\alias{fitted.values.trendfilter}
\alias{fitted.values.cv_trendfilter}
\alias{fitted.values.sure_trendfilter}
\alias{fitted.cv_trendfilter}
\alias{fitted.sure_trendfilter}
\title{Get fitted values from a trendfilter object}
\usage{
\method{fitted}{trendfilter}(obj, lambdas = NULL)
}
\arguments{
\item{obj}{Object of class \code{trendfilter}.}

\item{lambdas}{(Optional) Vector of lambda values to calculate fitted values at. If
missing, will use break points in the fit.}
}
\description{
Get fitted values from a trendfilter object
}
