% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sure_tf}
\alias{predict.sure_tf}
\title{Predict method for trend filtering fits}
\usage{
\method{predict}{sure_tf}(obj, lambda, x_eval, nx_eval)
}
\arguments{
\item{obj}{An object of class \code{\link[=sure_trendfilter]{sure_tf}} or
\code{\link[=cv_trendfilter]{cv_tf}}.}

\item{x_eval}{(Optional) Overrides \code{nx_eval} if passed. A grid of inputs to
evaluate the optimized trend filtering estimate on.}

\item{nx_eval}{Integer. The length of the input grid that the optimized
trend filtering estimate is evaluated on; i.e. if nothing is passed to
\code{x_eval}, then it is defined as
\code{x_eval = seq(min(x), max(x), length = nx_eval)}.}
}
\value{
A list with the following elements:
\describe{
\item{x_eval}{Input grid used to evaluate the optimized trend filtering
estimate on.}
\item{tf_estimate}{Optimized trend filtering estimate, evaluated at \code{x_eval}.}
\item{fitted_values}{Optimized trend filtering estimate, evaluated at the
observed inputs \code{x}.}
\item{residuals}{\code{residuals = y - fitted_values}}
}
}
\description{
Predict values based on a \code{\link[=sure_trendfilter]{sure_tf}} or
\code{\link[=cv_trendfilter]{cv_tf}} object.
}
