% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R, R/unfinished-scraps.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\alias{predict.cv_trendfilter}
\title{Get predictions from a trendfilter object}
\usage{
\method{predict}{trendfilter}(obj, lambda = NULL, x_eval = NULL, zero_tol = 1e-06, ...)

\method{predict}{cv_trendfilter}(
  obj,
  lambda = NULL,
  x_eval = NULL,
  loss_func = "MAE",
  zero_tol = 1e-06,
  ...
)
}
\arguments{
\item{obj}{An object of class '\code{\link[=trendfilter]{trendfilter}}'.}

\item{lambda}{One or more lambda values to evaluate predictions for. Defaults to
\code{lambda = NULL}, in which case, predictions are computed for every model in
\code{obj$lambda}.}

\item{x_eval}{Vector of inputs where the trend filtering model(s) will be evaluated.
Defaults to \code{x_eval = NULL}, in which case \code{x_eval = obj$x}.}

\item{zero_tol}{Threshold parameter that controls the point at which a small coefficient
value is set to zero. Defaults to \code{zero_tol = 1e-6}.}

\item{...}{Additional named arguments. Currently unused.}
}
\description{
Get predictions from a trendfilter object

Get predictions from a cv_trendfilter object
}
