% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{print.summary.trendfilter}
\alias{print.summary.trendfilter}
\alias{print.summary.cv_trendfilter}
\alias{print.summary.sure_trendfilter}
\title{Print the output of a trendfilter summary object}
\usage{
\method{print}{summary.trendfilter}(obj)
}
\arguments{
\item{obj}{Object of class \code{\link[=trendfilter]{trendfilter}}.}
}
\description{
Print the output of a trendfilter summary object
}
