% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{print.trendfilter}
\alias{print.trendfilter}
\title{Print the output of a trendfilter object}
\usage{
\method{print}{trendfilter}(obj)
}
\arguments{
\item{obj}{Object of class \code{\link[=trendfilter]{trendfilter}}.}
}
\description{
Print the output of a trendfilter object
}
