% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar_spec}
\alias{quasar_spec}
\title{Lyman-alpha forest in the absorption spectrum of a distant quasar}
\format{
A tibble
}
\source{
\url{http://www.sdss3.org/}
}
\usage{
quasar_spec
}
\description{
Lyman-alpha forest in the absorption spectrum of a distant quasar
}
\keyword{datasets}
