% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar_spectrum}
\alias{quasar_spectrum}
\title{Lyman-alpha forest in the absorption spectrum of a distant quasar}
\format{
A tibble with column set \code{c("log10_wavelength", "flux", "weights")}.
}
\source{
\url{http://www.sdss3.org/}
}
\usage{
quasar_spectrum
}
\description{
Lyman-alpha forest in the absorption spectrum of a distant quasar
}
\keyword{datasets}
