% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{residuals.trendfilter}
\alias{residuals.trendfilter}
\alias{residuals.cv_trendfilter}
\alias{residuals.sure_trendfilter}
\alias{resids.trendfilter}
\alias{resids.cv_trendfilter}
\alias{resids.sure_trendfilter}
\title{Get residuals from a trendfilter object}
\usage{
\method{residuals}{trendfilter}(obj, lambdas = NULL)
}
\arguments{
\item{obj}{Object of class \code{trendfilter}.}

\item{lambdas}{(Optional) Vector of lambda values to calculate residuals at. If
missing, will use break points in the fit.}
}
\description{
Get residuals from a trendfilter object
}
