% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sci}
\alias{sci}
\alias{var}
\alias{bpm}
\title{Sample SALT Observatory spectrum.}
\format{
Three list objects --- \code{sci}, \code{var}, \code{bpm}, --- that contain the
spectroscopic measurements, the measurement variances, and the pixel masks,
respectively, of the stellar spectrum.
}
\source{
\url{https://astronomers.salt.ac.za/data/download-data-and-calibrations/}
}
\usage{
sci
}
\description{
A stellar spectrum from the Southern African Large Telescope (SALT)
Observatory located in Karoo, South Africa.
}
\keyword{datasets}
