% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.trendfilter.R
\name{parametric.sampler}
\alias{parametric.sampler}
\alias{nonparametric.resampler}
\alias{wild.sampler}
\title{Resampling functions for various bootstrap algorithms}
\usage{
parametric.sampler(data)

nonparametric.resampler(data)

wild.sampler(data)
}
\arguments{
\item{data}{Data frame / tibble with minimal column set: \code{x}, \code{y}, \code{weights}
(for \code{parametric.sampler}), \code{fitted.values} (for \code{parametric.sampler}), and
\code{residuals} (for \code{wild.sampler})}
}
\value{
Bootstrap sample returned in the same format as the input data frame
/ tibble.
}
\description{
Resampling functions for various bootstrap algorithms
}
