% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sure_trendfilter.R
\name{sure_tf-class}
\alias{sure_tf-class}
\alias{sure_tf}
\title{\code{sure_tf} class}
\description{
A list object returned by \code{\link[=sure_trendfilter]{sure_trendfilter()}}, with the following elements:
\itemize{
\item{\code{x_eval}:} Input grid used to evaluate the optimized trend filtering
estimate on.
\item{\code{tf_estimate}:} Optimized trend filtering estimate, evaluated at
\code{x_eval}.
\item{\code{validation_method}:} "SURE"
\item{\code{lambdas}:} Vector of hyperparameter values evaluated in the grid search
(always returned in descending order).
\item{\code{edfs}:} Vector of effective degrees of freedom for all trend filtering
estimators fit during validation.
\item{\code{generalization_errors}:} Vector of SURE generalization error estimates,
corresponding to the descending-ordered \code{lambdas} vector.
\item{\code{lambda_min}:} Hyperparameter value that minimizes the SURE
generalization error curve.
\item{\code{i_min}:} Index of \code{lambdas} that minimizes the SURE error curve.
\item{\code{edf_min}:} Effective degrees of freedom of the optimized trend
filtering estimator.
\item{\code{n_iter}:} The number of iterations needed for the ADMM algorithm to
converge within the given tolerance, for each hyperparameter value. If many
of these are exactly equal to \code{max_iter}, then their solutions have not
converged with the tolerance specified by \code{obj_tol}. In which case, it is
often prudent to increase \code{max_iter}.
\item{\code{training_errors}:} Mean-squared error between the observed outputs \code{y}
and the trend filtering estimate, for every hyperparameter choice.
\item{\code{optimisms}:} SURE-estimated optimisms, i.e.
\code{optimisms = generalization_errors - training_errors}.
\item{\code{x}:} Vector of observed inputs.
\item{\code{y}:} Vector of observed outputs.
\item{\code{weights}:} Weights for the observed outputs, defined as the reciprocal
variance of the additive noise that contaminates the signal.
\item{\code{fitted_values}:} Optimized trend filtering estimate, evaluated at the
observed inputs \code{x}.
\item{\code{residuals}:} \code{residuals = y - fitted_values}
\item{\code{k}:} Degree of the trend filtering estimator.
\item{\code{admm_params}:} List of parameter settings for the trend filtering ADMM
algorithm, constructed by passing the \code{optimization_params} list to
\code{\link[glmgen:trendfilter.control.list]{glmgen::trendfilter.control.list()}}.
\item{\code{thinning}:} Logical. If \code{TRUE}, then the data are preprocessed so that a
smaller, better conditioned data set is used for fitting.
\item{\code{seed}:} Random number seed (for reproducible results).
\item{\code{x_scale}, \code{y_scale}, \code{data_scaled}:} For internal use.
}
}
\seealso{
\code{\link[=sure_trendfilter]{sure_trendfilter()}}
}
