% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilter.R
\name{trendfilter}
\alias{trendfilter}
\title{Fit a trend filtering model (back-end function with more options for expert
users)}
\usage{
trendfilter(x, y, weights = NULL, lambda, edf = NULL, ...)
}
\arguments{
\item{x}{Vector of observed values for the input variable.}

\item{y}{Vector of observed values for the output variable.}

\item{weights}{Weights for the output measurements. Output weights are defined as the
inverse variance of the additive noise that contaminates the output signal.
When the noise is expected to have a constant variance \mjseqn{\sigma^2}
over all outputs, a scalar may be passed to \code{weights}, i.e.
\verb{weights = }\mjseqn{1/\sigma^2}. Otherwise, \code{weights} must be a vector with
the same length as \code{x} and \code{y}.}

\item{lambda}{One or more hyperparameter values to fit a trend filtering estimate for.}

\item{edf}{(Not yet available) Alternative hyperparametrization for the trend
filtering model(s). Vector of the desired number of effective degrees of
freedom in each model.}

\item{...}{Additional named arguments to pass to the internal/expert function
\code{\link[=.trendfilter]{.trendfilter()}}.}
}
\value{
An object of class \code{'trendfilter'}.
}
\description{
Fit a trend filtering model. Generic functions such as \code{\link[=predict]{predict()}},
\code{\link[=fitted.values]{fitted.values()}}, and \code{\link[=residuals]{residuals()}} may be called on the
\code{\link[=trendfilter]{trendfilter()}} output.
}
\examples{
data("eclipsing_binary")
head(eclipsing_binary)

x <- eclipsing_binary$phase
y <- eclipsing_binary$flux
weights <- 1 / eclipsing_binary$std_err^2

fit <- trendfilter(x, y, weights, lambda = exp(10))
}
\references{
\enumerate{
\item Politsch et al. (2020a). Trend filtering – I. A modern statistical tool
for time-domain astronomy and astronomical spectroscopy. \emph{MNRAS}, 492(3),
p. 4005-4018.
[\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Publisher}]
[\href{https://arxiv.org/abs/1908.07151}{arXiv}].
\item Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical
signals with varying degrees of smoothness. \emph{MNRAS}, 492(3), p. 4019-4032.
[\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{Publisher}]
[\href{https://arxiv.org/abs/2001.03552}{arXiv}].
}
}
