% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbands.R
\name{vbands}
\alias{vbands}
\title{Compute bootstrap variability bands}
\usage{
vbands(obj, level = 0.95)
}
\arguments{
\item{obj}{An object of class \code{\link[=bootstrap_trendfilter]{'bootstrap_tf'}}.}

\item{level}{The level of the pointwise variability bands. Defaults to
\code{level = 0.95}.}
}
\value{
A tibble with column set:
\code{c("x","bootstrap_lower_band","bootstrap_upper_band")}.
}
\description{
Compute variability bands for the optimized trend filtering estimate via the
sample quantiles of the bootstrap ensemble generated by
\code{\link[=bootstrap_trendfilter]{bootstrap_trendfilter()}}.
}
\examples{
data(quasar_spectrum)
head(spec)

sure_tf <- sure_trendfilter(spec$log10_wavelength, spec$flux, spec$weights)
boot_tf <- bootstrap_trendfilter(sure_tf, bootstrap_algorithm = "parametric")
bands <- vbands(boot_tf)
}
\references{
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}
}
\seealso{
\code{\link[=bootstrap_trendfilter]{bootstrap_trendfilter()}}
}
