% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_trendfilter.R
\name{bootstrap_trendfilter}
\alias{bootstrap_trendfilter}
\title{Construct pointwise variability bands via a bootstrap}
\usage{
bootstrap_trendfilter(
  obj,
  bootstrap_algorithm = c("nonparametric", "parametric", "wild"),
  B = 100L,
  mc_cores = parallel::detectCores() - 4
)
}
\arguments{
\item{obj}{An object of class '\code{\link[=predict_trendfilter]{pred_tf}}'.}

\item{bootstrap_algorithm}{A string specifying which variation of the
bootstrap to use. One of \code{c("nonparametric", "parametric", "wild")}. See
details section below for when each option is appropriate.}

\item{B}{The number of bootstrap samples used to estimate the pointwise
variability bands. Defaults to \code{B = 100L}.}

\item{mc_cores}{Multi-core computing using the
\code{\link[parallel:parallel-package]{parallel}} package: The number of cores to
utilize. Defaults to the number of cores detected, minus 4.}
}
\value{
An object of class '\code{bootstrap_tf}'. This is a list with the
following elements
\describe{
\item{x_eval}{}
\item{ensemble}{The full trend filtering bootstrap ensemble as an
\mjseqn{n \times B} matrix.}
\item{edf_boots}{Vector of the estimated number of effective degrees of
freedom of each trend filtering bootstrap estimate. These should all be
relatively close to \code{obj$edf_min}.}
\item{n_iter_boots}{Vector of the number of iterations taken by the ADMM
algorithm before reaching a stopping criterion, for each bootstrap estimate.}
\item{lambda_boots}{Vector of the hyperparameter values used for each
bootstrap fit. In general, these are not all equal because our bootstrap
implementation instead seeks to hold the number of effective degrees of
freedom constant across all bootstrap estimates.}
\item{bootstrap_algorithm}{A string specifying which variation of the
bootstrap was used to generate the ensemble.}
}
}
\description{
\loadmathjax Generate a bootstrap ensemble of trend filtering estimates in
order to quantify the uncertainty in the optimized estimate. One of three
possible bootstrap algorithms should be chosen according to the criteria in
the details section below. See \href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Politsch et al. (2020a)} for the technical
details of each bootstrap algorithm. Pointwise variability bands are then
obtained by passing the \code{bootstrap_trendfilter()} output and a desired level
(e.g. \code{level = 0.95}) to \code{\link[=vbands]{vbands()}}.
}
\details{
Our recommendations for when to use each of the possible settings
for the \code{bootstrap_algorithm} argument are shown in the table below. See
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Politsch et al. (2020a)} for more details.\tabular{ll}{
   Scenario \tab Uncertainty quantification \cr
   \code{x} is unevenly sampled \tab \code{bootstrap_algorithm = "nonparametric"} \cr
   \code{x} is evenly sampled and reciprocal variances are available \tab \code{bootstrap_algorithm = "parametric"} \cr
   \code{x} is evenly sampled and reciprocal variances are not available \tab \code{bootstrap_algorithm = "wild"} \cr
}


For our purposes, an evenly sampled data set with some discarded pixels
(either sporadically or in large consecutive chunks) is still considered to
be evenly sampled. When the inputs are evenly sampled on a transformed scale,
we recommend transforming to that scale and carrying out the full trend
filtering analysis on that scale. See Example 2 below for a case when the
inputs are evenly sampled on the \code{log10(x)} scale.
}
\examples{
# Example 1: Phase-folded light curve of an eclipsing binary star system
#
# The apparent brightness over time of a star system that has two suns
# that regularly eclipse one another from our vantage point on Earth. Here,
# the time series is stacked according to the orbital period of the binary
# system, with the primary eclipse occuring at `phase = 0` and the input
# domain ranging from -0.5 to 0.5.

data(eclipsing_binary)
head(EB)

cv_tf <- cv_trendfilter(
  x = EB$phase,
  y = EB$flux,
  weights = 1 / EB$std_err^2,
  optimization_params = list(
    max_iter = 1e4,
    obj_tol = 1e-6,
    thinning = TRUE
  )
)

pred_tf <- predict(
  cv_tf,
  validation_error_metric = "MAE",
  lambda_choice = "lambda_1se",
  nx_eval = 1500L
)

boot_tf <- bootstrap_trendfilter(pred_tf, "nonparametric")
bands <- vbands(boot_tf)


# Example 2: The "Lyman-alpha forest" in the spectrum of a distant quasar

data(quasar_spectrum)
head(spec)

sure_tf <- sure_trendfilter(spec$log10_wavelength, spec$flux, spec$weights)
pred_tf <- predict(sure_tf)

boot_tf <- bootstrap_trendfilter(pred_tf, "parametric")
bands <- vbands(boot_tf)
}
\references{
\bold{Companion references}
\enumerate{
\item{Politsch et al. (2020a).
\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{
Trend filtering – I. A modern statistical tool for time-domain astronomy and
astronomical spectroscopy}. \emph{MNRAS}, 492(3), p. 4005-4018.} \cr
\item{Politsch et al. (2020b).
\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{
Trend Filtering – II. Denoising astronomical signals with varying degrees of
smoothness}. \emph{MNRAS}, 492(3), p. 4019-4032.}}

\bold{The Bootstrap and variations}
\enumerate{
\item{Efron and Tibshirani (1986).
\href{https://projecteuclid.org/journals/statistical-science/volume-1/issue-1/Bootstrap-Methods-for-Standard-Errors-Confidence-Intervals-and-Other-Measures/10.1214/ss/1177013815.full}{
Bootstrap Methods for Standard Errors, Confidence Intervals, and Other
Measures of Statistical Accuracy}.
\emph{Statistical Science}, 1(1), p. 54-75.} \cr
\item{Mammen (1993).
\href{https://projecteuclid.org/journals/annals-of-statistics/volume-21/issue-1/Bootstrap-and-Wild-Bootstrap-for-High-Dimensional-Linear-Models/10.1214/aos/1176349025.full}{
Bootstrap and Wild Bootstrap for High Dimensional Linear Models}. \emph{The
Annals of Statistics}, 21(1), p. 255-285.} \cr
\item{Efron (1979).
\href{https://projecteuclid.org/journals/annals-of-statistics/volume-7/issue-1/Bootstrap-Methods-Another-Look-at-the-Jackknife/10.1214/aos/1176344552.full}{
Bootstrap Methods: Another Look at the Jackknife}.
\emph{The Annals of Statistics}, 7(1), p. 1-26.}}
}
\seealso{
\code{\link[=cv_trendfilter]{cv_trendfilter()}}, \code{\link[=sure_trendfilter]{sure_trendfilter()}}
}
