% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{coef.trendfilter}
\alias{coef.trendfilter}
\alias{coef.cv_trendfilter}
\alias{coef.sure_trendfilter}
\title{Get coefficients from a trendfilter object}
\usage{
\method{coef}{trendfilter}(obj, lambdas = NULL)
}
\arguments{
\item{obj}{Object of class \code{trendfilter}.}

\item{lambdas}{(Optional) Vector of lambda values to calculate coefficients at. If
missing, will use break points in the fit.}
}
\description{
Get coefficients from a trendfilter object
}
