% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendfilter.R
\name{.trendfilter}
\alias{.trendfilter}
\alias{dot-trendfilter}
\title{Fit a trend filtering model}
\usage{
.trendfilter(
  x,
  y,
  weights = NULL,
  lambda,
  edfs = NULL,
  k = 2L,
  obj_tol = 1e-10,
  max_iter = length(y),
  ...
)
}
\arguments{
\item{x}{Vector of observed values for the input variable.}

\item{y}{Vector of observed values for the output variable.}

\item{weights}{Weights for the output measurements. Output weights are defined as the
inverse variance of the additive noise that contaminates the output signal.
When the noise is expected to have a constant variance \mjseqn{\sigma^2}
over all outputs, a scalar may be passed to \code{weights}, i.e.
\verb{weights = }\mjseqn{1/\sigma^2}. Otherwise, \code{weights} must be a vector
with the same length as \code{x} and \code{y}.}

\item{lambda}{Vector of one or more hyperparameter values to fit a model for.}

\item{edfs}{(Not yet available) Alternative hyperparametrization for trend filtering
model(s). Vector of the desired number of effective degrees of freedom in
each model.}

\item{k}{Degree of the polynomials that make up the piecewise-polynomial trend
filtering estimate. Defaults to \code{k = 2} (i.e. a piecewise quadratic
estimate). Must be one of \verb{k = 0,1,2}. Higher order polynomials are
disallowed since they yield no statistical benefit over \code{k = 2} and their
use can lead to instability in the convex optimization.}

\item{obj_tol}{Stopping criterion for the trend filtering convex optimization. If the
relative change in the trend filtering objective function between two
successive iterations is less than \code{obj_tol}, the algorithm terminates.
Defaults to \code{obj_tol = 1e-10}.}

\item{max_iter}{Maximum number of iterations that we will tolerate for the trend filtering
convex optimization algorithm. Defaults to \code{max_iter = length(y)}.}

\item{...}{Additional named arguments. Currently unused.}
}
\value{
An object of class \code{'trendfilter'}.
}
\description{
Fit a trend filtering model.
}
\references{
\enumerate{
\item Politsch et al. (2020a). Trend filtering – I. A modern statistical tool
for time-domain astronomy and astronomical spectroscopy. \emph{MNRAS}, 492(3),
p. 4005-4018.
[\href{https://academic.oup.com/mnras/article/492/3/4005/5704413}{Publisher}]
[\href{https://arxiv.org/abs/1908.07151}{arXiv}].
\item Politsch et al. (2020b). Trend Filtering – II. Denoising astronomical
signals with varying degrees of smoothness. \emph{MNRAS}, 492(3), p. 4019-4032.
[\href{https://academic.oup.com/mnras/article/492/3/4019/5704414}{Publisher}]
[\href{https://arxiv.org/abs/2001.03552}{arXiv}].
}
}
