% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{fitted.trendfilter}
\alias{fitted.trendfilter}
\alias{fitted.values.trendfilter}
\alias{fitted.values.cv_trendfilter}
\alias{fitted.values.sure_trendfilter}
\alias{fitted.values.bootstrap_trendfilter}
\alias{fitted.cv_trendfilter}
\alias{fitted.sure_trendfilter}
\alias{fitted.bootstrap_trendfilter}
\title{Fitted values of a trend filtering model}
\usage{
\method{fitted}{trendfilter}(object, ...)
}
\arguments{
\item{object}{Object of class "\code{\link[=trendfilter]{trendfilter}}".}

\item{...}{Additional named arguments.}
}
\description{
Get fitted values from a "\code{\link[=trendfilter]{trendfilter}}" object.
}
\details{
As of now, the additional named arguments that can be passed are:
\describe{
\item{\code{lambda}}{One or more lambda values to compute fitted values for.
By default, fitted values are computed for every model in
\code{object$lambda}.}
\item{\code{x_eval}}{Vector of inputs where the trend filtering model(s) will be
evaluated. By default, \code{x_eval = object$x}.}
\item{\code{zero_tol}}{(For experts) Threshold parameter that controls the point
at which a small coefficient value is set to zero. Defaults to
\code{zero_tol = 1e-10}.}
}
}
