% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sure_tf}
\alias{predict.sure_tf}
\title{Predict method for trend filtering fits}
\usage{
\method{predict}{sure_tf}(obj, lambda, x_eval, nx_eval)
}
\arguments{
\item{obj}{An object of class \code{\link[=sure_trendfilter]{sure_tf}} or
\code{\link[=cv_trendfilter]{cv_tf}}.}

\item{x_eval}{(Optional) Overrides \code{nx_eval} if passed. A grid of inputs to
evaluate the optimized trend filtering estimate on.}

\item{nx_eval}{Integer. The length of the input grid that the optimized
trend filtering estimate is evaluated on; i.e. if nothing is passed to
\code{x_eval}, then it is defined as
\code{x_eval = seq(min(x), max(x), length = nx_eval)}.}

\item{lambda_choice}{One of \code{c("lambda_min","lambda_1se")}. The choice
of hyperparameter that is used for optimized trend filtering estimate.
Defaults to \code{lambda_choice = "lambda_min"}.
\itemize{
\item{\code{"lambda_min"}}: The hyperparameter value that minimizes the cross
validation error curve.
\item{\code{"lambda_1se"}}: The largest hyperparameter value with a cross
validation error within 1 standard error of the minimum cross validation
error. This choice therefore favors simpler (i.e. smoother) trend filtering
estimates. The motivation here is essentially Occam's razor: the two models
yield results that are quantitatively very close, so we favor the simpler
model. See Section 7.10 of
\href{https://web.stanford.edu/~hastie/Papers/ESLII.pdf}{Hastie, Tibshirani, and Friedman (2009)}
for more details on the "one-standard-error rule".}}
}
\value{
A list with the following elements:
\describe{
\item{x_eval}{Input grid used to evaluate the optimized trend filtering
estimate on.}
\item{tf_estimate}{Optimized trend filtering estimate, evaluated at \code{x_eval}.}
\item{fitted_values}{Optimized trend filtering estimate, evaluated at the
observed inputs \code{x}.}
\item{residuals}{\code{residuals = y - fitted_values}}
}
}
\description{
Predict values based on a \code{\link[=sure_trendfilter]{sure_tf}} or
\code{\link[=cv_trendfilter]{cv_tf}} object.
}
