% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\title{Predict from a trend filtering model}
\usage{
\method{predict}{trendfilter}(object, ...)
}
\arguments{
\item{object}{An object of class/subclass '\code{\link[=trendfilter]{trendfilter}}'.}

\item{...}{Additional named arguments.}
}
\description{
Get predictions from a '\code{\link[=trendfilter]{trendfilter}}' object.
}
\details{
As of now, the additional parameters that can be passed are:
\describe{
\item{\code{lambda}}{One or more lambda values to evaluate predictions for.
By default, predictions are computed for every model in \code{object$lambda}.}
\item{\code{x_eval}}{Vector of inputs where the trend filtering model(s) will be
evaluated. By default, \code{x_eval = object$x}.}
\item{\code{zero_tol}}{(For experts) Threshold parameter that controls the point
at which a small coefficient value is set to zero. Defaults to
\code{zero_tol = 1e-10}.}
}
}
