% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\title{Get predictions from a trendfilter object}
\usage{
\method{predict}{trendfilter}(obj, lambdas, x_eval = NULL, zero_tol = 1e-06, ...)
}
\arguments{
\item{obj}{output of \code{\link[=trendfilter]{trendfilter()}}}

\item{lambdas}{vector of lambda values to calculate coefficients
at. If missing, will use break points in the fit.}

\item{x_eval}{vector of new x points. Set to NULL (the default) to use the
original locations.}

\item{zero_tol}{numerical tolerance parameter, for determining whether a}

\item{...}{optional, currently unused, arguments}
}
\description{
Get predictions from a trendfilter object
}
