% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{predict.trendfilter}
\alias{predict.trendfilter}
\title{Get predictions from a trendfilter object}
\usage{
\method{predict}{trendfilter}(obj, lambda = NULL, x_eval = NULL, zero_tol = 1e-10, ...)
}
\arguments{
\item{obj}{An object of class/subclass '\code{\link[=trendfilter]{trendfilter}}'.}

\item{lambda}{One or more lambda values to evaluate predictions for. Defaults to
\code{lambda = NULL}, in which case, predictions are computed for every model in
\code{obj$lambda}.}

\item{x_eval}{Vector of inputs where the trend filtering model(s) will be evaluated.
Defaults to \code{x_eval = NULL}, in which case \code{x_eval = obj$x}.}

\item{zero_tol}{Threshold parameter that controls the point at which a small coefficient
value is set to zero. Defaults to \code{zero_tol = 1e-6}.}

\item{...}{Additional named arguments. Currently unused.}
}
\description{
Get predictions from a '\code{\link[=trendfilter]{trendfilter}}' object
}
