% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{quasar_spectrum}
\alias{quasar_spectrum}
\title{Spectroscopic measurements of a distant quasar}
\format{
A tibble with column set \code{c("log10_wavelength", "flux", "weights")}.
}
\source{
\url{http://www.sdss3.org/}
}
\usage{
quasar_spectrum
}
\description{
Part of the electromagnetic spectrum of an extremely luminous galaxy. This
special type of galaxy -- a quasar -- is the most luminous object class in
the Universe. The specific interval of its spectrum that is studied here
traces the structure of diffuse gas in intergalactic space, along the quasar
light's one-dimensional path to Earth. The intergalactic gas that the light
passes through is primarily composed of neutrally-charged hydrogen, which
creates a "forest" of absorptions in this specific interval of the quasar's
spectrum.
}
\keyword{datasets}
