% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{residuals.trendfilter}
\alias{residuals.trendfilter}
\alias{residuals.cv_trendfilter}
\alias{residuals.sure_trendfilter}
\alias{residuals.bootstrap_trendfilter}
\alias{resids.trendfilter}
\alias{resids.cv_trendfilter}
\alias{resids.sure_trendfilter}
\alias{resids.bootstrap_trendfilter}
\title{Residuals of a trend filtering model}
\usage{
\method{residuals}{trendfilter}(object, ...)
}
\arguments{
\item{object}{Object of class "\code{\link[=trendfilter]{trendfilter}}".}

\item{...}{Additional named arguments.}
}
\description{
Get residuals from a "\code{\link[=trendfilter]{trendfilter}}" object.
}
\details{
As of now, the additional named arguments that may be passed are:
\describe{
\item{\code{lambda}}{One or more lambda values to compute model residuals for.
By default, residuals are computed for every model in
\code{object$lambda}.}
\item{\code{x_eval}}{Vector of inputs where the trend filtering model(s) will be
evaluated. By default, \code{x_eval = object$x}.}
\item{\code{zero_tol}}{(For experts) Threshold parameter that controls the point
at which a small coefficient value is set to zero. Defaults to
\code{zero_tol = 1e-10}.}
}
}
