% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-stats.R
\name{residuals.trendfilter}
\alias{residuals.trendfilter}
\alias{residuals.cv_trendfilter}
\alias{residuals.sure_trendfilter}
\alias{residuals.bootstrap_trendfilter}
\alias{resids.trendfilter}
\alias{resids.cv_trendfilter}
\alias{resids.sure_trendfilter}
\alias{resids.bootstrap_trendfilter}
\title{Residuals of a trend filtering model}
\usage{
\method{residuals}{trendfilter}(obj, lambda = NULL, x_eval = NULL, zero_tol = NULL, ...)
}
\arguments{
\item{obj}{Object of class \code{trendfilter}.}

\item{lambda}{One or more lambda values to compute residuals for. Defaults to
\code{lambda = NULL}, in which case, residuals are computed for all
hyperparameter values in \code{obj$lambda}.}

\item{x_eval}{Vector of inputs where the trend filtering model(s) will be evaluated.
Defaults to \code{x_eval = NULL}, in which case \code{x_eval = obj$x}.}

\item{zero_tol}{Threshold parameter that controls the point at which a small coefficient
value is set to zero. Defaults to \code{zero_tol = 1e-10}.}

\item{...}{Additional named arguments. Currently unused.}
}
\description{
Get residuals from a '\code{\link[=trendfilter]{trendfilter}}' object.
}
