% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-base.R
\name{summary.trendfilter}
\alias{summary.trendfilter}
\title{Summarize a trendfilter object}
\usage{
\method{summary}{trendfilter}(obj)
}
\arguments{
\item{obj}{Object of class \code{\link[=trendfilter]{trendfilter}}.}
}
\description{
Summarize a trendfilter object
}
